+++ src/vue/modules/coinpayments-deposit/module.js
@@ -3,7 +3,7 @@ import { ModuleDescriptor } from '@/modules-arch/module-descriptor'
 export class CoinpaymentsDepositModule extends ModuleDescriptor {
   constructor (opts = {}) {
     super({
-      importComponentFn: _ => import('@/vue/modules/coinpayments'),
+      importComponentFn: _ => import('@/vue/modules/coinpayments-deposit'),
       ...opts,
     })
   }